% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_nrows.R
\name{assert_nrows}
\alias{assert_nrows}
\title{Assert that a data.frame contains a specified number of rows}
\usage{
assert_nrows(data, target_nrows)
}
\arguments{
\item{data}{A data.frame or data.table}

\item{target_nrows}{Numeric -- number of expected rows}
}
\value{
Throws error if test is violated
}
\description{
Given a data.frame or data.table object and a target number of rows, check that a dataset has that many rows
}
\examples{
assert_nrows(CO2,84)
}

